File organization
Organized the files into specific locations. The contents of the deleted file are available in the GroundModeBranch.
