+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Robotics_10650_2024_2025_Code/TeleOp/captureTele.java
@@ -1,14 +1,12 @@
-package org.firstinspires.ftc.teamcode.Robotics_10650_2024_2025_Code.TextFiles;
+package org.firstinspires.ftc.teamcode.Robotics_10650_2024_2025_Code.TeleOp;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorEx;
 
-import org.firstinspires.ftc.teamcode.Robotics_10650_2024_2025_Code.RobotInitialize;
+import org.firstinspires.ftc.teamcode.Robotics_10650_2024_2025_Code.InitializeFolder.ProgBotInitialize;
+import org.firstinspires.ftc.teamcode.Robotics_10650_2024_2025_Code.TextFiles.ProgBotInitialize;
 
 import java.io.FileWriter;
-import java.io.IOException;
 import java.util.ArrayList;
 
     @TeleOp(name = "RecordDriverInputs", group = "Linear Opmode")
